refactory()
Includes the plugin in the global navigation and sitemap for discoverability.

Updates README to clearly state the project's always free and open source status.
