+++ gl-core/components/layout/RenderMarkdown.tsx
@@ -2,7 +2,6 @@
 
 import * as React from 'react';
 import ReactMarkdown from 'react-markdown';
-// import { Icon } from '../../gl-core';
 import { Box, Link as MuiLink, Typography, useTheme } from '@mui/material';
 
 export type TRenderMarkdown = {
