+++ public/sitemap.xml
@@ -29,6 +29,9 @@
     <loc>https://goldlabel.pro/cv</loc>
   </url>
   <url>
+    <loc>https://goldlabel.pro/flash</loc>
+  </url>
+  <url>
     <loc>https://goldlabel.pro/</loc>
   </url>
   <url>
